+++ idea/testData/codeInsight/overrideImplement/equalsInInterface.kt
@@ -1,0 +1,5 @@
+interface A
+
+interface B : A {
+    <caret>
+}